requirements.txt
pyhamcrest
assertpy
parameterized
coverage
nose2
coveralls